+++ packages/server/src/plugins/ws/utils/index.ts
@@ -3,5 +3,15 @@ import { updateCommonGames } from "./updateCommonGames";
 import { checkCommonGames } from "./checkCommonGames";
 import { deleteWaitlist } from "./deleteWaitlist";
 import { checkGameEnd } from "./checkGameEnd";
+import { fillPlayerGamesList } from "./fillPlayerGamesList";
+import { startWaitlist } from "./startWaitlist";
 
-export { calculateAllGames, updateCommonGames, checkCommonGames, deleteWaitlist, checkGameEnd };
+export {
+  calculateAllGames,
+  updateCommonGames,
+  checkCommonGames,
+  deleteWaitlist,
+  checkGameEnd,
+  fillPlayerGamesList,
+  startWaitlist,
+};
